booking

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->boolean('in_code')->default(0);
             $table->boolean('consent')->default(0);
              $table->string('name_admin')->nullable();
-            $table->integer('balance')->default(0);
+            $table->float('balance')->default(0);
             $table->string('profile_photo')->nullable();
 
             $table->rememberToken();
